+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [unreleased]
 
 - Support Compound borrow, repayBorrow and claimComp with DSProxy in `HSCompound` and `FCompoundActions`.
+- Support Stake, unstake and claimAll `$COMBO` in `HFurucombo`.
 
 ## [0.12.1] - 2020-10-28