+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+## [0.13.0] - 2020-12-08
+
+### Added
+
 - Support Compound borrow, repayBorrow and claimComp with DSProxy in `HSCompound` and `FCompoundActions`.
 - Support Stake, unstake and claimAll `$COMBO` in `HFurucombo`.